fix: Import works with handlebars templates
When importing data using handlebars templates, we had errors like `SyntaxError: Unexpected token { in JSON at position 31` because we were trying to read the doctypes from the JSON file before transforming it.
